+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-=======
-
 ## UNRELEASED
 
 - Update `examples/one-page.html` to `examples/js/index.js` with new script helper.
+- Add support for Deck or Slide-level transitions
+- Add default transitions for Fade, Slide, and None
 
 ## 6.0.0-alpha.7